auth folder reorganized
